+++ package.json
@@ -31,6 +31,7 @@
     "webpack-dev-server": "^3.11.0"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.10.4",
     "clean-webpack-plugin": "^3.0.0",
     "express": "^4.17.1",
     "html-webpack-plugin": "^4.3.0",
@@ -38,6 +39,7 @@
     "react-dom": "^16.13.1",
     "react-icons": "^3.9.0",
     "react-redux": "^7.2.0",
+    "react-router-dom": "^5.2.0",
     "react-transition-group": "^4.3.0",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
