+++ src/components/beamingDot/redBeamingDot.css
@@ -1,0 +1,144 @@
+
+.red-animation_main_div{
+    width:20px;
+  }
+  
+  
+  .red-circle {
+      width: 10px;
+      height: 10px;
+      border-radius: 50%;
+      background: #FAA;
+       position: absolute;
+      -webkit-animation-timing-function: linear;
+      -webkit-animation: scaler 4s infinite;  /* Chrome, Safari, Opera */
+      -webkit-animation-timing-function: linear;  /* Chrome, Safari, Opera */
+      animation: scaler 4s infinite;
+      animation-timing-function: linear;
+  }
+  
+  .red-circle-two {
+      width: 10px;
+      height: 10px;
+      border-radius: 50%;
+      background: #FAA;
+       position: absolute;
+      -webkit-animation-timing-function: linear;
+      -webkit-animation: scaler 4s infinite;  /* Chrome, Safari, Opera */
+      -webkit-animation-timing-function: linear;  /* Chrome, Safari, Opera */
+      animation: scaler 4s infinite;
+      animation-timing-function: linear;
+      -webkit-animation-delay: 1s;
+  }
+  
+  .red-circle-three {
+      width: 10px;
+      height: 10px;
+      border-radius: 50%;
+      background: #FAA;
+       position: absolute;
+      -webkit-animation-timing-function: linear;
+      -webkit-animation: scaler 4s infinite;  /* Chrome, Safari, Opera */
+      -webkit-animation-timing-function: linear;  /* Chrome, Safari, Opera */
+      animation: scaler 4s infinite;
+      animation-timing-function: linear;
+      -webkit-animation-delay: 2s;
+  }
+  
+  
+  .red-circle-four {
+      width: 10px;
+      height: 10px;
+      border-radius: 50%;
+      background: #FAA;
+       position: absolute;
+      -webkit-animation-timing-function: linear;
+      -webkit-animation: scaler 4s infinite;  /* Chrome, Safari, Opera */
+      -moz-animation: scaler 4s infinite;  /* Chrome, Safari, Opera */
+      animation: scaler 4s infinite;  /* Chrome, Safari, Opera */
+      -webkit-animation-timing-function: linear;  /* Chrome, Safari, Opera */
+      animation: scaler 4s infinite;
+      animation-timing-function: linear;
+      -webkit-animation-delay: 3s;
+  }
+  
+  
+  
+  @-webkit-keyframes scaler{
+  
+      0%{
+          -webkit-transform: scale(1);
+      }
+  
+      33%{
+          -webkit-transform: scale(4);
+          opacity: 0.20;
+      }
+  
+      67%{
+          -webkit-transform: scale(7);
+          opacity: 0.10;
+      }
+  
+      100%{
+          -webkit-transform: scale(10);
+          opacity: 0;
+      }
+  
+  
+  }
+  
+  @-moz-keyframes scaler{
+  
+      0%{
+          -moz-transform: scale(1);
+      }
+  
+      33%{
+          -moz-transform: scale(4);
+          opacity: 0.20;
+      }
+  
+      67%{
+          -moz-transform: scale(7);
+          opacity: 0.10;
+      }
+  
+      100%{
+          -moz-transform: scale(10);
+          opacity: 0;
+      }
+  
+  
+  }
+  
+  @keyframes scaler{
+  
+      0%{
+          transform: scale(1);
+      }
+  
+      33%{
+          transform: scale(4);
+          opacity: 0.20;
+      }
+  
+      67%{
+          transform: scale(7);
+          opacity: 0.10;
+      }
+  
+      100%{
+          transform: scale(10);
+          opacity: 0;
+      }
+  
+  
+  }
+  
+  
+  .holder {
+    display: flex;
+    justify-content: center;
+    padding-top: 30px;
+  }
