+++ src/components/responseHolder/ErrorLabel.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './errorLabel.css';
+
+export const ErrorLabel = ({text}) => {
+    return (
+        <label className="error-label">
+            {text}
+        </label>    
+    )
+}
