+++ src/components/responseHolder/FetchingLabel.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './fetchingLabel.css';
+
+export const FetchingLabel = ({text}) => {
+    return (
+        <label className="fetching-label">
+            {text}
+        </label>    
+    )
+}
