+++ src/components/responseHolder/badge.css
@@ -1,0 +1,12 @@
+#info {
+    background-color: #2196F3;
+    color: black;
+    opacity: 0.30;
+    border-radius: 5px;
+    padding: 5px;
+}
+
+#icon {
+    background-color: black;
+    border-radius: 5px;
+}
