+++ src/components/responseHolder/fetchingLabel.css
@@ -1,0 +1,27 @@
+@keyframes shimmer {
+    0%   {
+            background-color: palegoldenrod;
+            opacity: 0.30;
+        }
+    25%  {
+            background-color: palegoldenrod;
+            opacity: 0.20;
+        }
+    50%  {
+            background-color: palegoldenrod;
+            opacity: 0.30;
+        }
+    100% {
+            background-color: palegoldenrod;
+            opacity: 0.5;
+        }
+  }
+
+.fetching-label {
+    background-color: palegoldenrod;
+    color: black;
+    padding: 2px;
+    border-radius: 3px;
+    animation: shimmer 5s linear 2s infinite alternate;
+}
+
